+++ src/Traits/Comparison/ApproxComparable.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Galaxon\Core\Traits;
+namespace Galaxon\Core\Traits\Comparison;
 
 use Galaxon\Core\Exceptions\IncomparableTypesException;
 use Galaxon\Core\Floats;